apps/server-python/ASimpleStarGazer_python/requirements.txt
httpx>=0.28.1
python-dotenv>=1.0.0
mcp[cli]>=1.12.3